requirements.txt
flask>=3.0
flask-sqlalchemy>=3.1
apscheduler>=3.10
gunicorn>=21.2
python-dotenv>=1.0
